Separate functions and classes into modules
