+++ index.html
@@ -18,7 +18,7 @@
                 <li class="menu-item"><a href="#" id="disp-contact" class="link-item">Contact</a></li>
             </ul>
         </nav>
-        <span class="date-item">May 30th 2022</span>        
+        <span class="date">May 30th 2022</span>        
     </header>
     
     <main>
