+++ scripts/modules/book_class.js
@@ -1,0 +1,7 @@
+export default class Book {
+  constructor(title = null, author = null, id = null) {
+    this.title = title;
+    this.author = author;
+    this.id = id;
+  }
+}
