+++ scripts/modules/reset_input.js
@@ -1,0 +1,6 @@
+import {bookName, bookAuthor} from '../index.js';
+
+export default function resetInput() {
+    bookName.value = '';
+    bookAuthor.value = '';
+  }
